+++ repository/src/main/java/pl/inz/repository/CompanyRepository.java
@@ -5,15 +5,15 @@ import org.springframework.stereotype.Repository;
 import pl.inz.model.Company;
 
 @Repository
-public interface CompanyRepository extends CrudRepository<Company, Integer> {
+public interface CompanyRepository extends CrudRepository<Company, Long> {
 
     Iterable<Company> findAll();
 
-    Company findOne(Integer id);
+    Company findOne(Long id);
 
     Company save(Company newCompany);
 
-    void delete(Integer id);
+    void delete(Long id);
 
     long count();
 }
